- moved all the old commands out into a different file...just didn't want to keep seeing old errors in stuff that's going to be deleted anyway
